The Nc4Iosp code assumes that the netcdf C library nc_get/put_vars code works for user defined types. The netcdf-C documentations says it does, but it is not true because it uses nc_get/put_varm which only works for atomic types.
I have made a temporary fix. The nc_get/put_vara code
does handle user defined types. So I modified
the Nc4Iosp code so that if the stride is all ones,
then nc_get/put_vara is used instead of vars.

I have fixed the problem in the netcdf-c library
(see NCF-228) and as soon as it propagates
to the dlls used by Nc4Iosp, the temporary fix
can be removed.