Implement readTimeseries at ThreddsScalarLayer to improve efficiency
We have implemented the method readTimeseries at ThreddsScalarLayer in order to improve efficiency when accessing data values for a single point.
The problem is that the generic method defined at AbstractScalarLayer (in ncwms) calls readHorizontalPoints for each time included in the timeseries requested and there are some variables that could be calculated only once. With this new method, we avoid repeating this calculations more than one time (CdmUtils.createHorizontalGrid, for example).