Update LambertConformal.java
I have a sample dataset of WRF temperature data for the contiguous US that uses Lambert conformal conic gridding with a projection origin of latitude = +90 degrees. It cannot be opened by the NJ library because the LambertConformal class throws an exception.

The test in LambertConformal's precalculate() method whether lat0 equals +/- PI_OVER_2 is incorrect, as this is a possibly valid choice. I have altered the test to check instead that abs(lat0) is not greater than PI_OVER_2. Doing so allows NJ to open the dataset, and subsequently there is no mathematical error in any of the projection calculations.

The tests of whether the two standard parallels are valid are correct in trying to test whether those two values equal +/- 90 degrees (and if so reject them) but they did not set whether those values were outside the range (-90,90).