Added some mac related fixes to toolsUI
If on mac osx, use system default look and feel. This enables mac key bindings for copy paste. Fixes Unidata/thredds#163

If on a mac, make sure windows are closed when using quit from the system menu. This will trigger the exit method so that the preferences are written out just like if toolsUI was quit using the window 'x' button. [TDS-671]