Fix calls to setContentLength(). (Fixes #197)
setContentLength() sets the content-length HTTP header, which is defined
as the number of bytes. In all of these changed places, we were passing
in the length of the string, which is the number of characters. This
breaks in the presence UTF-8 characters, which may be encoded as up to 4
bytes; the result is truncated HTML data (observed in #197).