GINI: Clean-up zlib decompression code.
Refactor the zlib code to be clearer. Also, simplify by only trying to
grab the bytes for the decompressed data block, just ignoring the end of
file marker. This fixes an issue where the contract of Array.factory()
was broken by giving it more data than corresponds to the passed-in
shape.