Improve performance of timeseries reading.
The current implementation of timeseries reading in `ThreddsScalarLayer`
queries the dataset once for each involved time instant with
`CdmUtils.readHorizontalPoints`. The original layer implementation in
ncwms uses the function `CdmUtils.readTimeseries`, which retrieves
all the data from the earliest to the latest time instant at once.
With the modifications in socib/ncWMS@timeseries_fast_read_thredds
the loading can be delegated to that function with a significant
improvement in speed and caching opportunities. In addition,
this is the same approach used to implement `readHorizontalPoints`.
Please note that this will only work with the above mentioned changes to ncwms.

See Unidata/ncWMS#18 and Unidata/thredds#439.