Fix NEXRAD Level 2 IOSP to work with caching. (Fixes #435)
The underlying problem was readData using the raf from the internal
volScan object rather than the raf that gets properly initialized when
the netcdffile was reloaded from cache.

There is absolutely no reason for volScan to be a class attribute, so
convert it to a local variable, and pass it as needed to other
functions. This also eliminates a private internal function that was
making use of the attribute.