Fixes Unidata/thredds#470
Because we use ucar.unidata.util.Format.dfrac before writing out
the lat/lon bounding box on the dataset.html page for NCSS for Points,
there is an implicit "rounding" via truncation that happens. In the case
of a single station with a precise lat/lon location, the bounding box can
actually not contain the station. This fix corrects the effect of the
truncation based on how many decimal places you truncate to.