Link to artifacts on FTP server, not Nexus If you follow the release.txt document, it tells you to upload artifacts to the ftp server and then update some symlinks. If you do that, the NetCDF-Java and TDS front pages don't need to be updated.
For 5.0, I'd like to manage the links as properties in AsciiDoctor, which will cause them to be updated automatically. At that point, we should go back to pointing to Nexus.

In this commit, I also added more detailed instructions to release.txt about making release announcements. Up to 27 steps, oof! Must automate soon.