netCDFtools.jnlp is insulated from version changes
* All references to versioned artifacts (e.g. "ui-4.6.6.jar") have been moved to netCDFtoolsExtraJars.jnlp, which netCDFtools.jnlp loads as an extension.
* Both JNLP files use the "current" symlink in their codebase URL, instead of something like "v4.6".

The net effect of these changes is that as long as we're publishing our webstart artifacts to http://www.unidata.ucar.edu/software/thredds/current/netcdf-java/webstart,
netCDFtools.jnlp should "just work", even across version increments, and even if the file is downloaded and cached. In fact, there shouldn't be a need to change it for the
remainder of 4.X; only netCDFtoolsExtraJars.jnlp will need to be modified.

This fixes DYV-254597, although in an alternative way to what was suggested in the thread.