Fix conflicts from merging master into 5.0
Some of the APIs in 5.0 have diverged quite a bit, so this was mostly about old code being updated to use the new interface.

Most of the work was dedicated to resurrecting DsgSubsetWriterTest. It existed in 4.6 but had been deleted in 5.0.
The tests were DEEPLY broken by some of the API changes and–as far as I can tell–John just decided to delete the
file than to fix it. It's still a good test, so I've done the necessary work.

Also, made some changes to .travis.yml that should hopefully result in quicker builds. As part of this,
the :tds:compileGwt task will no longer be run, which is good as it was failing on Travis.