Catch NoSuchMethodError in CoverageDatasetFactory
When trying to load a netCDF file as via FeatureDatasetFactoryManager.open(...), the openGrib method thows a NoSuchMethodError. openGrib is called in the open method as a test to see if the dataset is a Grib file, and because the NoSuchMethodError wasn't being caught, the whole thing stops and the FeatureDatasetFactoryManager returns a null dataset with an error message in the Formatter. Adding this error to the catch allows FeatureDatasetFactoryManager to do the right thing and see that the file isn't a Grib file.

Because the error can be triggered by reflection, need to look at the target of a InvocationTargetException. Refactored the try...catch block to check both in one catch rather than dupicate code with two catch blocks.