Properly track whether files are opened/closed in Nc4Iosp (fixes #761)
Previously, after Nc4Iosp.close() was called, its "isClosed" flag would always be "true", even if open(), openForWriting(), or create() were later called. This causes a problem for subsequent calls of close() because the Nc4Iosp thinks it's *already* closed, and therefore doesn't invoke nc_close on the underlying file. And thus, the file is left open.

This issue reared its head when attempting to edit an existing NetCDF-4 file using NetcdfFileWriter on Windows. After the edit, the file could not be deleted, because it was still open. The file COULD be deleted on Linux and OSX. Given that the behavior only manifests on Windows, I decided not to write a regression test.