TDS Catalog pages: Render additional information from markdown if present
1. For all catalog pages render ${tds.content.path}/README.md if present.
2. For static catalogs e.g. ${tds.content.path}/subpath/example.xml render
${tds.content.path}/subpath/example.md
3. For dynamically generated catalogs render README.md in the directory that
was scanned for files, if that directory has no README.md, try in the parent
directory until reaching the parent catalog for the collection.

Markdown can be tested with a markdown editor (e.g.  https://stackedit.io/),
with github or similar.  Please note that HTML-tags are disable in this
implementation, which means that HTML-code that is embedded in Markdown file
will be esacped in the rendering process.