Replace use of 'provided' configuration with 'compileOnly'.
'provided' is added by "com.netflix.nebula:gradle-extra-configurations-plugin".
Since we began using it, the built-in Java plugin added 'compileOnly', which serves the same purpose.
So now, we can switch over and eliminate a dependency from the project.

Also, IntelliJ couldn't make sense of 'provided' and often failed to build the project.
It likes 'compileOnly' much better.