Replace the existing DAP4 .dmr Bison parser with a DOM-based parser.
Currently, the dap4 .dmr parser is written using Bison plus SAX.
Given that the standard .dmr format is XML, it seems more
reasonable to replace the existing Bison parser with a standard
DOM parser and operate by walking the DOM tree. This mirrors the
operation of the netcdf-c library.

There is a performance penalty because using Bison is very fast
compared to the standard org.w3c.dom DOM parser provided in the JVM.
The primary reason is that the Bison version does not need to create
the intermediate DOM tree.

On the plus side, I think the DOM version is significantly more
maintainable since Bison expertise is likely to be scarce in the future.
Note that the constraint language parser is still based on Bison.

The Bison parser code is left in the source tree for now until I
am sure there are not problems.

This change should affect only DAP4 and only a small part of that.