ENH: Allow joinNew NcML aggregations to use existing coordinates
For a joinNew aggregations, allow the use of an existing scalar variable
with the same name, and aggregate it. This allows, for instance, an
aggregation to create a new time dimension for CF-compliant files that
have a scalar coordinate variable for time (since there's only a single
time in the file).