Create only one MavenPublication per project.
Previously, webapp projects created two MavenPublications: 'java' and 'web', each with their own POM. When I tried to publish both to the new Nexus repo, the second publication was refused because I was trying to overwrite the first POM and repository policy forbids redeployment. Now, all artifacts for a project are included in one, big MavenPublication.

Updated tests to account for the change.