Improve TestNcDump
The line endings of expectedOutput and actualOutput are converted to Unix EOLs before the 2 are compared. This fixes a test failure on Windows.

Added testNestedGroups, which asserts that NcDump's indentation is correct for a complex, nested structure.

Tweaked the printing of newlines a bit. It's more consistent and prettier now.