Make special attributes accessible by name only. Re: git issue https://github.com/Unidata/thredds/issues/948
Currently, special attributes like _NCProperties are visible
when iterating thru the attributes of the root group.
This pr changes that so that they are suppressed when using
Group.getAttributes(). They can still be accessed by name using
Group.findAttribute(String name).
This change enforces the similar rule used in the netcdf-C library.

Added a test dataset and a test case for this change.