re: e-support ZIZ-818863 re: gihub pr https://github.com/Unidata/thredds/pull/955
The user has asked for the ability to use a larger output buffer
size for the stream to which the inflated file is written.  This
is a follow-on to the previous request to allow the user to set
the inflate buffer size.

The solution we have arrived at is to set the output buffer size
to be the same size as 8*chunk size being inflated (which is as
large or larger than the deflated size).  This means that the
chunk size for the input file will also determine the output
buffer size.

This fix is in master, if it works out, then I will add it to 5.0.0.
The fix only changes code in H5tiledLayoutBB.java, method
Datachunk#inflate().