Add new test class for logical views in NcML. Nuke old test class.
TestLogicalReduce has been failing for a while. This is a BAD class for several reasons:
* It depends on an external resource.
* It opens the test file as a FeatureDatasetPoint, but NcML processing happens at the NetcdfDataset level. Why involve the PointFeature machinery when it's not necessary for the test? It's just another potential source of bugs/failures.
* It makes assertions about the data that aren't true, namely that times are strictly increasing. The test was failing because that's not the case. But in this context, why do we even care about that?
* It provides nothing more than a smoke test of the "logicalReduce" functionality.

Basically, it was worse than useless. So I nuked it (plus supporting resources). In its place I added LogicalViewsSpec (plus supporting resources), which actually tests logical views in a sensible way.

I also fixed a bug in the NcML schema. In the "logicalSlice" element, I renamed the "dimNames" attribute to "dimName".