Rewrite NCSS front-end.
The new NCSS pages mostly look and act like the old pages, but the implementation is completely different. The markup, CSS, and Javascript are all brand new. Almost none of the old stuff survived.

We're now using Thymeleaf templates for markup generation. We already had one template for Grid, but now we have ones for Grid-as-Point, Point, and Station as well. I moved several common fragments of markup to their own templates. Those fragments then get included in the main pages. As a result, there is very little duplication of markup.

We're now using modern CSS3 for all of our page layout. No more tables or floating elements; CSS Flexbox and Grid all the way! Note that Flexbox and Grid are recent technologies and are unsupported in older browsers. CSS Grid in particular was not supported in all major browsers until Nov 2017. Perhaps I shouldn't have chosen to use such bleeding-edge tech, but layout would have been much harder and kludgy-er without it. Meh. People will just have to update their shit.

We're now using ECMAScript 6 WITHOUT jQuery. OpenLayers.js is still a dependency, and unfortunately we remain stuck on the 2.X line (4.X is the new stuff). The new Javascript largely does the same things that the old Javascript did, but in about a quarter of the code, due to better reuse of functions. The one big new feature is the accordion buttons on the Grid and Grid-as-Point pages.