Handle NaNs in nearlyEquals() and fix tests that use it.
Almost all test failures were due to a maxRelDiff that was too small.

A notable exception is TestCoverageCurvilinear.TestGribCurvilinearSubset(). It creates a subset of a coverage that fits within a provided lat/lon bounding box. Then, it takes the shape of the subset and compares it to an expected value. The expected shape was {1, 166, 160}, but we were getting {1, 165, 161}. I cannot account for the change; it seems unrelated to my new nearlyEquals() implementation. To fix the test, I've simply updated the expected shape, because it still correctly corresponds to the given lat/lon box.

Also, use Misc's relative difference constants in MAMath.nearlyEquals().