Calculate correct boundaries for coverages that cross longitude seam.
Added HorizCoordSys.calcConnectedLatLonBoundaryPoints(), which calculates that lat/lon boundary of the CRS.
* It has optional parameters that limit the resolution.
* It works for projection, latLon1D, and latLon2D CRSs.
* Its boundary points are "connected", meaning that longitude values of adjacent elements are adjusted to eliminate discontinuities resulting from crossing the seam. Consequently, when cross-seam boundaries are turned into WKT polygons and plotted in NCSS's OpenLayers map, they'll be properly interpreted.
* Similar code used to live in CoverageBoundariesExtractor, but it was incomplete and/or buggy. I nuked all the old stuff.

Rewrote HorizCoordSys.calcLatLonBoundingBox() to properly handle coverages that straddle the seam.
* It builds upon calcConnectedLatLonBoundaryPoints(). In fact, its implementation is quite simple once we have that.
* HorizCoordSys2D used to have its own implementation, but the HorizCoordSys version now handles all CRSs properly. Nuked it.

Added extensive Javadocs for the new/updated boundary methods.
* It was quite a tricky problem, as the docs will show. The ultimate solution turned out to be relatively simple, but it took a while for me to get there (and realize that it worked).

Added HorizCoordSysCrossSeamBoundarySpec, which extensively tests the boundary methods.
* It builds HorizCoordSys objects from the new "crossSeam*.ncml" test resources.

Did some other minor refactoring in HorizCoordSys, and updated the classes that depend on it.
* Renamed getIsProjection() to isProjection().
* Renamed makeProjectionBB() to calcProjectionBoundingBox().
* Renamed makeLatLonBB() to calcLatLonBoundingBox().
* Converted axis field names to camelCase.

DtCoverageAdapter now produces 1D longitude axes with "connected" values. It appears that we used to do that in the analogous dt.grid stack, but that the adjustment was lost in the transition to coverage. Hopefully this change doesn't break anything.

Revert changes to LatLonPointImpl that I made in last commit; I didn't consider the longitude -180 == 180 case.