Update all catalogs and docs to use NCSS's 5.0.0 service base
Basically, instead of:
       <service name="ncss" serviceType="NetCDFSubset" base="/thredds/ncss/" />

Now use one or both of:
       <service name="ncssGrid" serviceType="NetcdfSubset" base="/thredds/ncss/grid/" />
       <service name="ncssPoint" serviceType="NetcdfSubset" base="/thredds/ncss/point/" />

A lot of the global compound services (usually named "all" in the root catalogs) include both, because often you don't know the types of data that they'll be applied to ahead of time (could be GRID or POINT).

Updated unit tests to account for the additional service that was added to many of the test catalogs.