Publish to raw repo using multipart POST requests, not individual PUTs
Source files are packaged into a series of POST requests of roughly 5 MB. For small files--e.g. our Javadoc sets--this means that potentially hundreds of files will be sent to Nexus in the same request. This has shown dramatic speed increases for publishing "javadocAll" when compared to making one PUT request per file.

We're now publishing using the OkHttp HTTP client, not Java's built-in client. We needed this for the multipart POST requests. It's also allowed us to set the Authorization header in a simpler fashion.