Add custom assertions for comparing floating-point numbers
Previously, we were using:
    Assert.assertTrue(Misc.nearlyEquals(expected, actual));

The problem with that is it gives no feedback about the values when it fails: all it gives you is "false". In order to learn the expected/actual values that failed, you'd have to print them yourself or walk through the test in a debugger. The custom assertions automatically print those values in the event of a failure.

There are custom assertions for floats, doubles, arrays of floats, and arrays of doubles. They borrow some code from JUnit, so I took the proper steps for attribution. I also updated various READMEs to gel with the reorganization of the docs/ module.

In our test suite, I changed all instances of the old floating-point-comparison pattern to the new assertions. While I was in those test classes, I also converted STDOUT messages to SLF4J and performed various other cleanup. TestProjectTiming and TestXMLStoreChains were providing no value whatsoever, so I nuked them.