The focus here was two-fold.
First was getting TestServlet to work under Intellij using
Spring+Mocking.  I attempted a variety of approaches to get this
work, but ultimately gave up because the Thredds configuration
code fails in ChronicleMap when executing under Intellij.  So, I
gave up and used a hack as described in Dap4Controller.java.

Second was to experiment with using a lock around the Nc4Prototype
code to guarantee serialized access. This was because I was suspcious
that the Native.synchronizedLibrary operation might not be working
correctly and was causing the HDF5 library to fail. However, looking
at the Jenkins runs, I believe that Native.synchronizedLibrary is
working correctly. We shall see.