fix ncss page
In xsl land, empty elements are output without closing tags when the output is html.
This causes all kinds of issues with the table layout.
By inserting a <span></span> within the empty <div>s used for holding ncss map previews, we keep the right table nesting, and we don't get a collision of the url builder row and the right column widgets.
Thank you @ddirks for helping debug this madness!