Enable NCSS returns when variable has _ChunkSizes attribute
Currently, if an NCSS request is made to a dataset for a variable with
the _ChunkSizes attribute exposed, and the request is made for a
netCDF-4 return, the write will fail with a "Bad Chunk Sizes" error.
This has been properly fixed in netCDF-Java 5. This PR simply removes
the inproperly exposed attribute before writing of netCDF-4 files after
subset is attempted.