Add gradle dependency-check plugin
Port of Unidata/netcdf-java#500.

The OWAPS dependency-check plugin for Jenkins is up for adoption, and
given that it has been failing to run for a week or so, does not seem to
be the way to go forward. This PR adds a gradle plugin as a replacement.
We will continue to use Jenkins to run the check on a regular schedule,
but now using the gradle machinery in place of the jenkins plugin.

To run the check locally, use:

./gradlew dependencyCheckAggregate

The plugin will generate an html report which can be found at

build/reports/dependency-check-report.html