Migrate away from bintray and jcenter
Both the bintray and jcenter maven repositories have been sunsetted. We
now run a jcenter proxy on the Unidata artifacts server to ensure we can
still access the shadow plugin that v4.x relies on. Other dependencies
that we used to rely on jcenter for can be found on maven central (at
least it seems). However, the fork of gretty does not, and was publised
to the 3rd-party repository on the Unidata artifacts server.
For more information on the sunset of bintray and jcenter, see:
https://jfrog.com/blog/into-the-sunset-bintray-jcenter-gocenter-and-chartcenter/